fix(Configuration & Maintenance): NEPT-509 - Foward port of the branch nept-509 into the 2.4 dev branch
